Registration form loksins!

+++ user/models.py
@@ -10,7 +10,7 @@ class Country(models.Model):
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     profile_image = models.CharField(max_length=9999)
-    kennitala = models.IntegerField()
+    kennitala = models.BigIntegerField()
     phone_number = models.IntegerField()
     address = models.CharField(max_length=255)
     postal_code = models.ForeignKey('estate.Municipality', on_delete=models.CASCADE)
